text2video/requirements.txt
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
torch>=2.1.0
moviepy>=1.0.3
Pillow>=10.0.0
tqdm>=4.66.0
